Split dashboard ops helpers modules
